add pingnode to utils
