+++ packages/grid_client/scripts/applications/casberlabs.ts
@@ -1,18 +1,6 @@
 import { FilterOptions, GatewayNameModel, MachinesModel } from "../../src";
 import { config, getClient } from "../client_loader";
-import { log } from "../utils";
-
-async function pingNodes(client, nodes) {
-  for (const node of nodes) {
-    try {
-      await client.zos.pingNode({ nodeId: node.nodeId });
-      return node.nodeId;
-    } catch (error) {
-      throw new Error("node " + node.nodeId + " is not responding, trying different node.");
-    }
-  }
-  throw new Error("No avaiable nodes");
-}
+import { log, pingNodes } from "../utils";
 
 async function deploy(client, vms, subdomain, gatewayNode) {
   const resultVM = await client.machines.deploy(vms);
@@ -63,7 +51,7 @@ async function main() {
   const vmQueryOptions: FilterOptions = {
     cru: 2,
     mru: 4,
-    sru: 50,
+    sru: 100,
     availableFor: grid3.twinId,
     farmId: 1,
   };
@@ -87,7 +75,7 @@ async function main() {
         disks: [
           {
             name: "wedDisk",
-            size: 50,
+            size: 100,
             mountpoint: "/data",
           },
         ],
